Move out VH wrapping logic to other class
