+++ library/src/main/java/com/friendoye/recyclerxray/internal/ScannableRecyclerAdapter.kt
@@ -5,8 +5,6 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.FrameLayout
 import androidx.annotation.Dimension
-import androidx.constraintlayout.widget.ConstraintLayout
-import androidx.constraintlayout.widget.ConstraintSet
 import androidx.core.view.children
 import androidx.core.view.doOnLayout
 import androidx.core.view.isVisible
@@ -14,7 +12,6 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.friendoye.recyclerxray.*
 import com.friendoye.recyclerxray.testing.ExceptionShooter
-import kotlinx.android.synthetic.main.xray_item_debug_layout.view.*
 
 
 internal class ScannableRecyclerAdapter<T : RecyclerView.ViewHolder>(
@@ -30,6 +27,11 @@ internal class ScannableRecyclerAdapter<T : RecyclerView.ViewHolder>(
     private val scanner: Scanner = Scanner()
 ) : DelegateRecyclerAdapter<T>(decoratedAdapter),
     XRayCustomParamsAdapterProvider {
+
+    private val holderViewWrapper = HolderViewWrapper(
+        xRayDebugViewHolder,
+        minDebugViewSize
+    )
 
     private val innerAdapterWatcher = InnerAdapterWatcher<T>(
         xRayApi,
@@ -87,7 +89,9 @@ internal class ScannableRecyclerAdapter<T : RecyclerView.ViewHolder>(
             innerAdapterWatcher.startWatching(holder)
         }
 
-        val xRayContainer = wrap(holderWrapper, holderItemParams)
+        val xRayContainer = holderViewWrapper.wrap(
+            holderWrapper, holderItemParams, currentOrientation
+        )
 
         return holder.replaceItemView(
             xRayWrapperContainer = xRayContainer,
@@ -149,54 +153,6 @@ internal class ScannableRecyclerAdapter<T : RecyclerView.ViewHolder>(
             showInnerAdapterIndicator = shouldShowInnerAdapterXRay,
             customParamsFromAdapter = provideCustomParams(holder.bindingAdapterPosition)
         )
-    }
-
-    private fun wrap(holderWrapperView: View, holderItemParams: ViewGroup.LayoutParams?): ConstraintLayout {
-        val context = holderWrapperView.context
-        val xRayContainer = ConstraintLayout(context).apply {
-            id = R.id.parent_constraint_layout_id
-            if (holderItemParams != null) {
-                layoutParams = holderItemParams
-            }
-        }
-
-        fun createMatchViewConstraint(childId: Int, parentId: Int) = ConstraintSet().apply {
-            connect(childId, ConstraintSet.LEFT,   parentId, ConstraintSet.LEFT)
-            connect(childId, ConstraintSet.RIGHT,  parentId, ConstraintSet.RIGHT)
-            connect(childId, ConstraintSet.TOP,    parentId, ConstraintSet.TOP)
-            connect(childId, ConstraintSet.BOTTOM, parentId, ConstraintSet.BOTTOM)
-            minDebugViewSize?.let {
-                when (currentOrientation) {
-                    RecyclerView.VERTICAL   -> constrainMinHeight(childId, it)
-                    RecyclerView.HORIZONTAL -> constrainMinWidth(childId, it)
-                }
-            }
-        }
-
-        xRayContainer.addView(holderWrapperView,
-            ConstraintLayout.LayoutParams.MATCH_PARENT, ConstraintLayout.LayoutParams.MATCH_PARENT
-        )
-
-        val debugLayout = xRayDebugViewHolder.provideView(xRayContainer).apply {
-            id = R.id.debug_layout_id
-            visibility = View.GONE
-            isClickable = true
-            tag = "DEBUG"
-        }
-        xRayContainer.addView(debugLayout)
-        val debugLayoutConstraints = createMatchViewConstraint(debugLayout.id, holderWrapperView.id)
-        debugLayoutConstraints.applyTo(xRayContainer)
-
-        val innerRecyclerIndicatorView = View(context).apply {
-            id = R.id.inner_indicator_view_id
-            visibility = View.GONE
-            setBackgroundResource(R.drawable.bg_inner_recycler_indicator)
-        }
-        xRayContainer.addView(innerRecyclerIndicatorView)
-        val indicatorViewConstraints = createMatchViewConstraint(innerRecyclerIndicatorView.id, holderWrapperView.id)
-        indicatorViewConstraints.applyTo(xRayContainer)
-
-        return xRayContainer
     }
 
     // TODO: Unstable API
